story-gs > staging: refactored
