--- src/bin/games/genius_square/models/arrays/mod.rs
@@ -1,5 +1,0 @@
-/// Models for handling arrays
-
-pub mod models;
-
-pub use models::BinArray;
