+++ src/bin/games/genius_square/models/binary_arrays/mod.rs
@@ -1,0 +1,7 @@
+/// Models for handling arrays
+
+pub mod models_arrays;
+pub mod models_grids;
+
+pub use models_arrays::BinArray;
+pub use models_grids::BinGrid;
