+++ src/bin/games/genius_square/models/binary_arrays/models_arrays.rs
